Update README.md
Hey there! I've been working on an E-commerce app using React, Redux, axios, and Sass. It's been a challenging project, but I'm excited about the results so far. The app allows users to browse a variety of products and make purchases online. I've been focusing on making the user experience as smooth and intuitive as possible. It's great to see how much React and Redux can simplify the process of building complex web applications. Thanks for taking the time to check out my work!